+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -31,6 +31,7 @@ CHANGELOG
  * Added the `messenger:setup-transports` command to setup messenger transports
  * Added a `InMemoryTransport` to Messenger. Use it with a DSN starting with `in-memory://`.
  * Added `framework.property_access.throw_exception_on_invalid_property_path` config option.
+ * Added `cache:pool:list` command to list all available cache pools.
 
 4.2.0
 -----